bot/requirements.txt
python-telegram-bot==13.1
tweepy==3.9.0
vk_api==11.9.1
